Feature/gat-4603

+++ app/Http/Controllers/Api/V1/ToolController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api\V1;
 use Config;
 use Auditor;
 use Exception;
+use App\Exceptions\NotFoundException;
 use App\Models\Tag;
 use App\Models\Tool;
 use App\Models\Dataset;
@@ -28,6 +29,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Carbon;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Tool\GetTool;
 use App\Http\Requests\Tool\EditTool;
@@ -142,6 +144,7 @@ class ToolController extends Controller
             $teamId = $request->query('team_id', null);
             $userId = $request->query('user_id', null);
             $filterTitle = $request->query('title', null);
+            $filterStatus = $request->query('status', null);
             $perPage = request('per_page', Config::get('constants.per_page'));
 
             $sort = $request->query('sort', 'name:desc');
@@ -186,6 +189,10 @@ class ToolController extends Controller
             ->when($userId, function ($query) use ($userId) {
                 return $query->where('user_id', '=', $userId);
             })
+            ->when($filterStatus, 
+            function ($query) use ($filterStatus) {
+                return $query->where('status', '=', $filterStatus);
+            })
             ->when($filterTitle, function ($query) use ($filterTitle) {
                 return $query->where('name', 'like', '%' . $filterTitle . '%');
             })
@@ -208,6 +215,74 @@ class ToolController extends Controller
             return response()->json(
                 $tools
             );
+        } catch (Exception $e) {
+            throw new Exception($e->getMessage());
+        }
+    }
+
+    /**
+     * @OA\Get(
+     *    path="/api/v1/tools/count/{field}",
+     *    operationId="count_unique_fields_tools",
+     *    tags={"Tools"},
+     *    summary="ToolController@count",
+     *    description="Get Counts for distinct entries of a field in the model",
+     *    security={{"bearerAuth":{}}},
+     *    @OA\Parameter(
+     *       name="field",
+     *       in="path",
+     *       description="name of the field to perform a count on",
+     *       required=true,
+     *       example="status",
+     *       @OA\Schema(
+     *          type="string",
+     *          description="status field",
+     *       ),
+     *    ),
+     *    @OA\Parameter(
+     *       name="team_id",
+     *       in="query",
+     *       description="team id",
+     *       required=true,
+     *       example="1",
+     *       @OA\Schema(
+     *          type="integer",
+     *          description="team id",
+     *       ),
+     *    ),
+     *    @OA\Response(
+     *       response="200",
+     *       description="Success response",
+     *       @OA\JsonContent(
+     *          @OA\Property(
+     *             property="data",
+     *             type="object",
+     *          )
+     *       )
+     *    )
+     * )
+     */
+    public function count(Request $request, string $field): JsonResponse
+    {
+        try {
+            $teamId = $request->query('team_id',null);
+            $counts = Tool::when($teamId, function ($query) use ($teamId) {
+                return $query->where('team_id', '=', $teamId);
+            })->withTrashed()
+                ->select($field)
+                ->get()
+                ->groupBy($field)
+                ->map->count();
+
+            Auditor::log([
+                'action_type' => 'GET',
+                'action_name' => class_basename($this) . '@'.__FUNCTION__,
+                'description' => "Tool count",
+            ]);
+
+            return response()->json([
+                "data" => $counts
+            ]);
         } catch (Exception $e) {
             throw new Exception($e->getMessage());
         }
@@ -311,6 +386,7 @@ class ToolController extends Controller
      *             @OA\Property( property="durs", type="array", @OA\Items() ),
      *             @OA\Property( property="collections", type="array", @OA\Items() ),
      *             @OA\Property( property="any_dataset", type="boolean", example=false ),
+     *             @OA\Property( property="status", type="string", enum={"ACTIVE", "DRAFT", "ARCHIVED"} ),
      *          ),
      *       ),
      *    ),
@@ -374,52 +450,55 @@ class ToolController extends Controller
                 'enabled',
                 'team_id', 
                 'mongo_id',
-                'associated_authors', 
+                'associated_authors',
                 'contact_address',
-                'any_dataset'
+                'any_dataset',
+                'status',
             ];
 
             $array = $this->checkEditArray($input, $arrayKeys);
             $tool = Tool::create($array);
+            $toolId = $tool->id;
 
-            $this->insertToolHasTag($input['tag'], (int) $tool->id);
+            $this->insertToolHasTag($input['tag'], (int) $toolId);
             if (array_key_exists('dataset', $input)) {
-                $this->insertDatasetVersionHasTool($input['dataset'], (int) $tool->id);
+                $this->insertDatasetVersionHasTool($input['dataset'], (int) $toolId);
             }
             if (array_key_exists('programming_language', $input)) {
-                $this->insertToolHasProgrammingLanguage($input['programming_language'], (int) $tool->id);
+                $this->insertToolHasProgrammingLanguage($input['programming_language'], (int) $toolId);
             }
             if (array_key_exists('programming_package', $input)) {
-                $this->insertToolHasProgrammingPackage($input['programming_package'], (int) $tool->id);
+                $this->insertToolHasProgrammingPackage($input['programming_package'], (int) $toolId);
             }
             if (array_key_exists('type_category', $input)) {
-                $this->insertToolHasTypeCategory($input['type_category'], (int) $tool->id);
+                $this->insertToolHasTypeCategory($input['type_category'], (int) $toolId);
             }
 
             $publications = array_key_exists('publications', $input) ? $input['publications'] : [];
-            $this->checkPublications($tool->id, $publications, $array['user_id'], $appId);
+            $this->checkPublications($toolId, $publications, $array['user_id'], $appId);
 
             if (array_key_exists('durs', $input)) {
-                $this->insertDurHasTool($input['durs'], (int) $tool->id);
+                $this->insertDurHasTool($input['durs'], (int) $toolId);
             }
 
             $collections = array_key_exists('collections', $input) ? $input['collections'] : [];
-            $this->checkCollections($tool->id, $collections, $array['user_id'], $appId);
+            $this->checkCollections($toolId, $collections, $array['user_id'], $appId);
 
-            if($request['enabled']){
-                $this->indexElasticTools((int) $tool->id);
+            $currentTool = Tool::where('id', $toolId)->first();
+            if($currentTool->status === Tool::STATUS_ACTIVE){
+                $this->indexElasticTools((int) $toolId);
             }
             
             Auditor::log([
                 'user_id' => (int) $jwtUser['id'],
                 'action_type' => 'CREATE',
                 'action_name' => class_basename($this) . '@'.__FUNCTION__,
-                'description' => "Tool " . $tool->id . " created",
+                'description' => "Tool " . $toolId . " created",
             ]);
 
             return response()->json([
                 'message' => 'created',
-                'data' => $tool->id,
+                'data' => $toolId,
             ], 201);
         } catch (Exception $e) {
             throw new Exception($e->getMessage());
@@ -472,12 +551,13 @@ class ToolController extends Controller
      *             @OA\Property( property="durs", type="array", @OA\Items() ),
      *             @OA\Property( property="collections", type="array", @OA\Items() ),
      *             @OA\Property( property="any_dataset", type="boolean", example=false ),
+     *             @OA\Property( property="status", type="string", enum={"ACTIVE", "DRAFT", "ARCHIVED"} ),
      *          ),
      *       ),
      *    ),
      *      @OA\Response(
-     *          response=201,
-     *          description="Created",
+     *          response=200,
+     *          description="Success",
      *          @OA\JsonContent(
      *              @OA\Property(property="message", type="string", example="success"),
      *              @OA\Property(property="data", type="integer", example="100")
@@ -511,6 +591,11 @@ class ToolController extends Controller
         try {
             $input = $request->all();
             $jwtUser = array_key_exists('jwt_user', $input) ? $input['jwt_user'] : [];
+            $initTool = Tool::withTrashed()->where('id', $id)->first();
+
+            if ($initTool['status'] === Tool::STATUS_ARCHIVED && !array_key_exists('status', $input)) {
+                throw new Exception('Cannot update current putoolblication! Status already "ARCHIVED"');
+            }
 
             $userId = null;
             $appId = null;
@@ -539,6 +624,7 @@ class ToolController extends Controller
                 'associated_authors', 
                 'contact_address',
                 'any_dataset',
+                'status',
             ];
 
             $array = $this->checkEditArray($input, $arrayKeys);
@@ -576,7 +662,8 @@ class ToolController extends Controller
             $collections = array_key_exists('collections', $input) ? $input['collections'] : [];
             $this->checkCollections($id, $collections, $array['user_id'], $appId);
 
-            if($request['enabled']){
+            $currentTool = Tool::where('id', $id)->first();
+            if($currentTool->status === Tool::STATUS_ACTIVE){
                 $this->indexElasticTools((int) $id);
             }
 
@@ -651,12 +738,13 @@ class ToolController extends Controller
      *             @OA\Property( property="durs", type="array", @OA\Items() ),
      *             @OA\Property( property="collections", type="array", @OA\Items() ),
      *             @OA\Property( property="any_dataset", type="boolean", example=false ),
+     *             @OA\Property( property="status", type="string", enum={"ACTIVE", "DRAFT", "ARCHIVED"} ),
      *          ),
      *       ),
      *    ),
      *      @OA\Response(
-     *          response=201,
-     *          description="Created",
+     *          response=200,
+     *          description="Success",
      *          @OA\JsonContent(
      *              @OA\Property(property="message", type="string", example="success"),
      *              @OA\Property(property="data", type="integer", example="100")
@@ -693,26 +781,33 @@ class ToolController extends Controller
 
             if ($request->has('unarchive')) {
                 // Restore the tool and related models
-                Tool::withTrashed()->where('id', $id)->restore();
-                ToolHasTag::withTrashed()->where('tool_id', $id)->restore();
-                DatasetVersionHasTool::withTrashed()->where('tool_id', $id)->restore();
-                ToolHasProgrammingLanguage::withTrashed()->where('tool_id', $id)->restore();
-                ToolHasProgrammingPackage::withTrashed()->where('tool_id', $id)->restore();
-                ToolHasTypeCategory::withTrashed()->where('tool_id', $id)->restore();
-                PublicationHasTool::withTrashed()->where('tool_id', $id)->restore();
-                DurHasTool::withTrashed()->where('tool_id', $id)->restore();
-                CollectionHasTool::withTrashed()->where('tool_id', $id)->restore();
+                $toolModel = Tool::withTrashed()
+                    ->find($id);
+                if (($request['status'] !== Tool::STATUS_ARCHIVED) && (in_array($request['status'], [
+                    Tool::STATUS_ACTIVE, Tool::STATUS_DRAFT
+                ]))) {
+                    $toolModel->status = $request['status'];
+                    $toolModel->deleted_at = null;
+                    $toolModel->save();
 
-                if ($request['enabled']) {
-                    $this->indexElasticTools($id);
+                    ToolHasTag::withTrashed()->where('tool_id', $id)->restore();
+                    DatasetVersionHasTool::withTrashed()->where('tool_id', $id)->restore();
+                    ToolHasProgrammingLanguage::withTrashed()->where('tool_id', $id)->restore();
+                    ToolHasProgrammingPackage::withTrashed()->where('tool_id', $id)->restore();
+                    ToolHasTypeCategory::withTrashed()->where('tool_id', $id)->restore();
+                    PublicationHasTool::withTrashed()->where('tool_id', $id)->restore();
+                    DurHasTool::withTrashed()->where('tool_id', $id)->restore();
+                    CollectionHasTool::withTrashed()->where('tool_id', $id)->restore();
+
+                    Auditor::log([
+                        'user_id' => (int) $jwtUser['id'],
+                        'action_type' => 'UPDATE',
+                        'action_name' => class_basename($this) . '@'.__FUNCTION__,
+                        'description' => "Tool " . $id . " unarchived",
+                    ]);
+                } else {
+                    throw new Exception('Cannot unarchive current tool because valid status not supplied');
                 }
-
-                Auditor::log([
-                    'user_id' => (int) $jwtUser['id'],
-                    'action_type' => 'UPDATE',
-                    'action_name' => class_basename($this) . '@'.__FUNCTION__,
-                    'description' => "Tool " . $id . " unarchived",
-                ]);
     
             }
             
@@ -741,9 +836,16 @@ class ToolController extends Controller
                 'associated_authors', 
                 'contact_address',
                 'any_dataset',
+                'status',
             ];
 
             $array = $this->checkEditArray($input, $arrayKeys);
+
+            $initTool = Tool::withTrashed()->where('id', $id)->first();
+
+            if ($initTool['status'] === Tool::STATUS_ARCHIVED && !array_key_exists('status', $input)) {
+                throw new Exception('Cannot update current tool! Status already "ARCHIVED"');
+            }
 
             Tool::where('id', $id)->update($array);
 
@@ -785,7 +887,8 @@ class ToolController extends Controller
                 $this->checkCollections($id, $collections, $userIdFinal, $appId);
             }
             
-            if ($request['enabled']) {
+            $currentTool = Tool::where('id', $id)->first();
+            if($currentTool->status === Tool::STATUS_ACTIVE){
                 $this->indexElasticTools($id);
             }
 
@@ -860,26 +963,33 @@ class ToolController extends Controller
             $input = $request->all();
             $jwtUser = array_key_exists('jwt_user', $input) ? $input['jwt_user'] : [];
 
-            Tool::where('id', $id)->delete();
-            ToolHasTag::where('tool_id', $id)->delete();
-            DatasetVersionHasTool::where('tool_id', $id)->delete();
-            ToolHasProgrammingLanguage::where('tool_id', $id)->delete();
-            ToolHasProgrammingPackage::where('tool_id', $id)->delete();
-            ToolHasTypeCategory::where('tool_id', $id)->delete();
-            PublicationHasTool::where('tool_id', $id)->delete();
-            DurHasTool::where('tool_id', $id)->delete();
-            CollectionHasTool::where('tool_id', $id)->delete();
-            
-            Auditor::log([
-                'user_id' => (int) $jwtUser['id'],
-                'action_type' => 'DELETE',
-                'action_name' => class_basename($this) . '@'.__FUNCTION__,
-                'description' => "Tool " . $id . " deleted",
-            ]);
+            $tool = Tool::where(['id' => $id])->first();
+            if ($tool) {
+                $tool->deleted_at = Carbon::now();
+                $tool->status = Tool::STATUS_ARCHIVED;
+                $tool->save();
+                ToolHasTag::where('tool_id', $id)->delete();
+                DatasetVersionHasTool::where('tool_id', $id)->delete();
+                ToolHasProgrammingLanguage::where('tool_id', $id)->delete();
+                ToolHasProgrammingPackage::where('tool_id', $id)->delete();
+                ToolHasTypeCategory::where('tool_id', $id)->delete();
+                PublicationHasTool::where('tool_id', $id)->delete();
+                DurHasTool::where('tool_id', $id)->delete();
+                CollectionHasTool::where('tool_id', $id)->delete();
+                
+                Auditor::log([
+                    'user_id' => (int) $jwtUser['id'],
+                    'action_type' => 'DELETE',
+                    'action_name' => class_basename($this) . '@'.__FUNCTION__,
+                    'description' => "Tool " . $id . " deleted",
+                ]);
 
-            return response()->json([
-                'message' => Config::get('statuscodes.STATUS_OK.message'),
-            ], Config::get('statuscodes.STATUS_OK.code'));
+                return response()->json([
+                    'message' => Config::get('statuscodes.STATUS_OK.message'),
+                ], Config::get('statuscodes.STATUS_OK.code'));
+            }
+
+            throw new NotFoundException();
         } catch (Exception $e) {
             throw new Exception($e->getMessage());
         }
@@ -898,9 +1008,10 @@ class ToolController extends Controller
             'publications',
             'durs',
             'collections',
-        ])->where([
-            'id' => $toolId,
-        ])->first();
+        ])
+        ->withTrashed()
+        ->where(['id' => $toolId])
+        ->first();
 
         $tool->setAttribute('datasets', $tool->allDatasets  ?? []);
         return $tool;
